refactor: review -- expand opossum model to multiple files
