+++ tests/core/entities/generators/opossum_provider.py
@@ -5,7 +5,8 @@ from typing import Any
 
 from faker.providers import BaseProvider
 
-from opossum_lib.core.entities.opossum import Opossum, ScanResults
+from opossum_lib.core.entities.opossum import Opossum
+from opossum_lib.core.entities.scan_results import ScanResults
 from opossum_lib.shared.entities.opossum_output_file_model import OpossumOutputFileModel
 from tests.core.entities.generators.scan_results_provider import ScanResultsProvider
 from tests.input_formats.opossum.entities.generators.generate_outfile_information import (  # noqa: E501
